+++ src/main/java/com/pml/dto/StorageDeviceDTO.java
@@ -22,7 +22,7 @@ public class StorageDeviceDTO implements Serializable {
 	private String description;
 	private boolean itWorks; 
 	private boolean itComposed;
-	private Integer sizeInGB;
+	private Double sizeInGB;
 	private Integer architecture;
 	private Integer type;
 	
@@ -126,11 +126,11 @@ public class StorageDeviceDTO implements Serializable {
 		this.itComposed = isItComposed;
 	}
 
-	public Integer getSizeInGB() {
+	public Double getSizeInGB() {
 		return sizeInGB;
 	}
 
-	public void setSizeInGB(Integer sizeInGB) {
+	public void setSizeInGB(Double sizeInGB) {
 		this.sizeInGB = sizeInGB;
 	}
